apps/ml/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12
pillow==10.4.0
numpy==1.26.4
onnxruntime==1.19.2
insightface==0.7.3
opencv-python-headless==4.10.0.84
playwright==1.48.0
pytesseract==0.3.13
psycopg[binary]==3.2.3
python-dotenv==1.0.1
